# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_upload_file": {
        "properties": {
          "document_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Добавить как новую ревизию существующего документа",
            "title": "Document Id"
          },
          "file": {
            "contentMediaType": "application/octet-stream",
            "description": "Файл проекта",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_file",
        "type": "object"
      },
...
      "DocumentKind": {
        "description": "Природа документа. BIM-форматы принимаются на хранение, но не разбираются.",
        "enum": [
          "pdf",
          "recognized_package",
          "revit",
          "navisworks",
          "ifc",
          "other"
        ],
        "title": "DocumentKind",
        "type": "string"
      },
      "DocumentRead": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "discipline": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Discipline"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "document_kind": {
            "$ref": "#/components/schemas/DocumentKind"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "project_id": {
            "format": "uuid",
            "title": "Project Id",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "project_id",
          "display_name",
          "discipline",
          "document_kind",
          "created_at",
          "updated_at"
        ],
        "title": "DocumentRead",
        "type": "object"
      },
      "DocumentRevisionRead": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "document_id": {
            "format": "uuid",
            "title": "Document Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "processing_error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Processing Error Code"
          },
          "processing_status": {
            "$ref": "#/components/schemas/ProcessingStatus"
          },
          "revision_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revision Label"
          },
          "source_filename": {
            "title": "Source Filename",
            "type": "string"
          },
          "source_metadata": {
            "additionalProperties": true,
            "title": "Source Metadata",
            "type": "object"
          },
          "source_mime": {
            "title": "Source Mime",
            "type": "string"
          },
          "source_sha256": {
            "title": "Source Sha256",
            "type": "string"
          },
          "source_size": {
            "title": "Source Size",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "document_id",
          "revision_label",
          "source_filename",
          "source_mime",
          "source_size",
          "source_sha256",
          "processing_status",
          "processing_error_code",
          "source_metadata",
          "created_at"
        ],
        "title": "DocumentRevisionRead",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "JobRead": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "job_type": {
            "$ref": "#/components/schemas/JobType"
          },
          "progress": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Progress"
          },
          "project_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          },
          "stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          }
        },
        "required": [
          "id",
          "project_id",
          "job_type",
          "status",
          "progress",
          "stage",
          "error_code",
          "error_message",
          "created_at",
          "started_at",
          "finished_at"
        ],
        "title": "JobRead",
        "type": "object"
      },
      "JobStatus": {
        "enum": [
          "queued",
          "running",
          "succeeded",
          "failed",
          "cancelled"
        ],
        "title": "JobStatus",
        "type": "string"
      },
      "JobType": {
        "description": "Типы заданий.\n\nРеален только legacy_import. Остальные объявлены в промте 08 как контракт и в Stage 1\nне исполняются.",
        "enum": [
          "legacy_import"
        ],
        "title": "JobType",
        "type": "string"
      },
...
      "ProcessingStatus": {
        "description": "Состояние обработки конкретной ревизии.\n\nЕдинственное изменяемое поле ревизии: описывает не файл, а ход работы над ним.",
        "enum": [
          "pending",
          "unprocessed",
          "processor_unavailable",
          "importing",
          "ready",
          "failed"
        ],
        "title": "ProcessingStatus",
        "type": "string"
      },
...
      "ProjectJobSummary": {
        "description": "Состояние последнего задания проекта.\n\nКомпактнее полного JobRead: список проектов не должен тащить служебные поля,\nкоторые в нём всё равно не показать.",
        "properties": {
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "job_type": {
            "$ref": "#/components/schemas/JobType"
          },
          "progress": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Progress"
          },
          "stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          }
        },
        "required": [
          "id",
          "job_type",
          "status",
          "progress",
          "stage",
          "error_code",
          "finished_at"
        ],
        "title": "ProjectJobSummary",
        "type": "object"
      },
      "ProjectRead": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "external_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Ref"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "source": {
            "$ref": "#/components/schemas/ProjectSource"
          },
          "status": {
            "$ref": "#/components/schemas/ProjectStatus"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "status",
          "source",
          "external_ref",
          "created_at",
          "updated_at"
        ],
        "title": "ProjectRead",
        "type": "object"
      },
...
      "ProjectSource": {
        "description": "Откуда взялся проект.\n\nНужен, чтобы отличить заведённый руками проект от подтянутого из внешней системы:\nу второго есть чужой идентификатор, и повторно создавать его нельзя.",
        "enum": [
          "manual",
          "tenderhub"
        ],
        "title": "ProjectSource",
        "type": "string"
      },
      "ProjectStatus": {
        "description": "Жизненный цикл проекта. Ход импорта сюда не пишется — он живёт в Job и ревизиях.",
        "enum": [
          "active",
          "archived"
        ],
        "title": "ProjectStatus",
        "type": "string"
      },
      "ProjectSummary": {
        "description": "Проект вместе с производными счётчиками — то, что нужно списку и карточке.\n\nПоследнее задание отдаётся здесь же: иначе, чтобы увидеть ход импорта, пришлось бы\nоткрывать каждый проект по очереди.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "document_count": {
            "minimum": 0.0,
            "title": "Document Count",
            "type": "integer"
          },
          "external_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Ref"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "last_job": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProjectJobSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "sheet_count": {
            "minimum": 0.0,
            "title": "Sheet Count",
            "type": "integer"
          },
          "source": {
            "$ref": "#/components/schemas/ProjectSource"
          },
          "status": {
            "$ref": "#/components/schemas/ProjectStatus"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "status",
          "source",
          "external_ref",
          "created_at",
          "updated_at",
          "document_count",
          "sheet_count"
        ],
        "title": "ProjectSummary",
        "type": "object"
      },
      "ProjectUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProjectStatus"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "ProjectUpdate",
        "type": "object"
      },
...
      "UploadRead": {
        "description": "Результат загрузки.\n\nis_duplicate означает, что такой файл в проекте уже был: возвращается прежняя ревизия,\nвторая копия не создаётся.",
        "properties": {
          "document": {
            "$ref": "#/components/schemas/DocumentRead"
          },
          "is_duplicate": {
            "title": "Is Duplicate",
            "type": "boolean"
          },
          "job": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "revision": {
            "$ref": "#/components/schemas/DocumentRevisionRead"
          }
        },
        "required": [
          "document",
          "revision",
          "job",
          "is_duplicate"
        ],
        "title": "UploadRead",
        "type": "object"
      },
      "UploadedFileType": {
        "description": "Что портал умеет делать с файлом такого расширения.",
        "properties": {
          "capability": {
            "title": "Capability",
            "type": "string"
          },
          "document_kind": {
            "$ref": "#/components/schemas/DocumentKind"
          },
          "extension": {
            "title": "Extension",
            "type": "string"
          },
          "max_size_bytes": {
            "exclusiveMinimum": 0.0,
            "title": "Max Size Bytes",
            "type": "integer"
          },
          "schedules_import": {
            "title": "Schedules Import",
            "type": "boolean"
          }
        },
        "required": [
          "extension",
          "document_kind",
          "capability",
          "schedules_import",
          "max_size_bytes"
        ],
        "title": "UploadedFileType",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "API портала автоматизированного подсчёта строительных объёмов. Stage 1: проекты, импорт распознанных пакетов и просмотрщик документации. Расчёта объёмов и вызовов моделей на этом этапе нет.",
    "title": "Quantor API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/projects/{project_id}": {
      "get": {
        "operationId": "read_project",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Проект",
        "tags": [
          "projects"
        ]
      },
      "patch": {
        "operationId": "update_project",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Изменить проект",
        "tags": [
          "projects"
        ]
      }
    },
...
    "/api/v1/projects/{project_id}/upload-capabilities": {
      "get": {
        "description": "Список поддерживаемых расширений и того, что портал с ними сделает.\n\nИнтерфейс берёт подписи отсюда, а не хранит свою копию: иначе при подключении нового\nобработчика обещания на экране разойдутся с поведением сервера.",
        "operationId": "read_upload_capabilities",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UploadedFileType"
                  },
                  "title": "Response Read Upload Capabilities",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Какие файлы принимает портал",
        "tags": [
          "uploads"
        ]
      }
    },
    "/api/v1/projects/{project_id}/uploads": {
      "post": {
        "description": "Принимает файл, кладёт его в хранилище и заводит документ с ревизией.\n\nРаспознавание здесь не выполняется: для ZIP-пакета создаётся задание импорта в состоянии\n«в очереди», для остальных типов ревизия сохраняется с честным статусом обработки.",
        "operationId": "upload_file",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_file"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Загрузить файл в проект",
        "tags": [
          "uploads"
        ]
      }
    },
...
  }
}